requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-rest-swagger==2.2.0
django-test-without-migrations==0.6
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.13
oauthlib==2.1.0
openapi-codec==1.3.2
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
uritemplate==3.0.0
urllib3==1.24.1
